support ~use_rgb parameter to filter points with RGB color
