+++ pr2_navigation_self_filter/src/self_filter.cpp
@@ -40,7 +40,6 @@
 #include <tf/message_filter.h>
 #include <message_filters/subscriber.h>
 
-    
 class SelfFilter
 {
 public:
@@ -48,15 +47,29 @@ public:
   SelfFilter(void): nh_("~")
   {
     nh_.param<std::string>("sensor_frame", sensor_frame_, std::string());
-    self_filter_ = new filters::SelfFilter<robot_self_filter::PointCloud>(nh_);
+    nh_.param("use_rgb", use_rgb_, false);
+    if (use_rgb_) 
+    {
+      self_filter_rgb_ = new filters::SelfFilter<pcl::PointXYZRGB>(nh_);
+    }
+    else 
+    {
+      self_filter_ = new filters::SelfFilter<pcl::PointXYZ>(nh_);
+    }
 
     sub_ = new message_filters::Subscriber<sensor_msgs::PointCloud2>(root_handle_, "cloud_in", 1);	
     mn_ = new tf::MessageFilter<sensor_msgs::PointCloud2>(*sub_, tf_, "", 1);
 
-    //mn_ = new tf::MessageNotifier<robot_self_filter::PointCloud>(tf_, boost::bind(&SelfFilter::cloudCallback, this, _1), "cloud_in", "", 1);
     pointCloudPublisher_ = root_handle_.advertise<sensor_msgs::PointCloud2>("cloud_out", 1);
     std::vector<std::string> frames;
-    self_filter_->getSelfMask()->getLinkNames(frames);
+    if (use_rgb_) 
+    {
+      self_filter_rgb_->getSelfMask()->getLinkNames(frames);
+    }
+    else 
+    {
+      self_filter_->getSelfMask()->getLinkNames(frames);
+    }
     if(frames.empty())
     {
       ROS_DEBUG("No valid frames have been passed into the self filter. Using a callback that will just forward scans on.");
@@ -72,7 +85,14 @@ public:
     
   ~SelfFilter(void)
   {
-    delete self_filter_;
+    if (self_filter_) 
+    {
+      delete self_filter_;
+    }
+    if (self_filter_rgb_)
+    {
+      delete self_filter_rgb_;
+    }
     delete mn_;
     delete sub_;
   }
@@ -84,27 +104,43 @@ private:
     ROS_DEBUG("Self filter publishing unfiltered frame");
   }
     
-  void cloudCallback(const sensor_msgs::PointCloud2::ConstPtr &cloud2)
+  void cloudCallback(const sensor_msgs::PointCloud2::ConstPtr &cloud2) 
   {
-    robot_self_filter::PointCloud out;
-      
     ROS_DEBUG("Got pointcloud that is %f seconds old", (ros::Time::now() - cloud2->header.stamp).toSec());
     std::vector<int> mask;
     ros::WallTime tm = ros::WallTime::now();
 
-    pcl::PointCloud<pcl::PointXYZ>::Ptr cloud(new pcl::PointCloud<pcl::PointXYZ>);
-    pcl::fromROSMsg(*cloud2, *cloud);
-
-    self_filter_->updateWithSensorFrame(*cloud, out, sensor_frame_);
+    
+    sensor_msgs::PointCloud2 out2;
+    int input_size = 0;
+    int output_size = 0;
+    if (use_rgb_)
+    {
+      typename pcl::PointCloud<pcl::PointXYZRGB>::Ptr cloud(new pcl::PointCloud<pcl::PointXYZRGB>);
+      pcl::fromROSMsg(*cloud2, *cloud);
+      pcl::PointCloud<pcl::PointXYZRGB> out;
+      self_filter_rgb_->updateWithSensorFrame(*cloud, out, sensor_frame_);
+      pcl::toROSMsg(out, out2);
+      input_size = cloud->points.size();
+      output_size = out.points.size();
+    }
+    else
+    {
+      typename pcl::PointCloud<pcl::PointXYZ>::Ptr cloud(new pcl::PointCloud<pcl::PointXYZ>);
+      pcl::fromROSMsg(*cloud2, *cloud);
+      pcl::PointCloud<pcl::PointXYZ> out;
+      self_filter_->updateWithSensorFrame(*cloud, out, sensor_frame_);
+      pcl::toROSMsg(out, out2);
+      input_size = cloud->points.size();
+      output_size = out.points.size();
+    }
       
     double sec = (ros::WallTime::now() - tm).toSec();
-
-    sensor_msgs::PointCloud2 out2;
-    pcl::toROSMsg(out, out2);
     pointCloudPublisher_.publish(out2);
-    ROS_DEBUG("Self filter: reduced %d points to %d points in %f seconds", (int)cloud->points.size(), (int)out.points.size(), sec);
-  }
+    ROS_DEBUG("Self filter: reduced %d points to %d points in %f seconds", input_size, output_size, sec);
 
+  }
+  
   tf::TransformListener                                 tf_;
   //tf::MessageNotifier<robot_self_filter::PointCloud>           *mn_;
   ros::NodeHandle                                       nh_, root_handle_;
@@ -112,9 +148,11 @@ private:
   tf::MessageFilter<sensor_msgs::PointCloud2>           *mn_;
   message_filters::Subscriber<sensor_msgs::PointCloud2> *sub_;
 
-  filters::SelfFilter<robot_self_filter::PointCloud> *self_filter_;
+  filters::SelfFilter<pcl::PointXYZ> *self_filter_;
+  filters::SelfFilter<pcl::PointXYZRGB> *self_filter_rgb_;
   std::string sensor_frame_;
-
+  bool use_rgb_;
+  
   ros::Publisher                                        pointCloudPublisher_;
   ros::Subscriber                                       no_filter_sub_;
   
@@ -124,7 +162,8 @@ private:
 int main(int argc, char **argv)
 {
   ros::init(argc, argv, "self_filter");
-
+  
+  ros::NodeHandle nh("~");
   SelfFilter s;
   ros::spin();
